Remove state machine gem from Spree::Shipment
The state machine gem is something that we don't want around long term,
as it has been known to cause headaches. The state machine in
`Spree::Shipment` is fairly straight forward, which makes it a good
candidate to try and remove. By removing the state machine gem, the
logic for states is now entirely contained within `Spree::Shipment`.
This helps the readability of the code and also makes it easier to
override behaviour that was previously buried within the state machine
gem.